Move data table versioning components into workspace-data
